Update dcos_metadata plugin to honor nested container IDs properly
Prevously, we were replacing the container ID of any nested containers
with the container ID of their parent as the index into the metrics
cache. This essentially meant that all nested containers were rolled up
under the same container ID of their parent in the metrics cache. This
commit fixes this.